fix: harden oauth and preheat workflows

+++ src-tauri/src/models.rs
@@ -124,7 +124,7 @@ pub struct CreditsSnapshot {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct RateLimitWindow {
-    pub used_percent: i32,
+    pub used_percent: f64,
     pub resets_at: Option<i64>,
     pub window_duration_mins: Option<i64>,
 }
